refactor: move GmatCommand class to new gmat_command.py
