+++ gmatpyplus/command/__init__.py
@@ -1,4 +1,5 @@
 from .branch_command import *
 from .command import *
+from .gmat_command import GmatCommand
 from .solver_branch_command import *
 from .solver_sequence_command import *
